Split Bob,java into OOP parts
